+++ admin/class-admin-asset-manager.php
@@ -226,9 +226,6 @@ class WPSEO_Admin_Asset_Manager {
 	 * @return array The scripts that need to be registered.
 	 */
 	protected function scripts_to_be_registered() {
-		$flat_version = $this->flatten_version( WPSEO_VERSION );
-		$ext_length   = ( strlen( $flat_version ) + 4 );
-
 		$header_scripts          = [
 			'admin-global',
 			'block-editor',
@@ -281,7 +278,7 @@ class WPSEO_Admin_Asset_Manager {
 		$plugin_scripts   = $this->load_generated_asset_file(
 			[
 				'asset_file'      => __DIR__ . '/../src/generated/assets/plugin.php',
-				'ext_length'      => $ext_length,
+				'ext_length'      => 3,
 				'additional_deps' => $additional_dependencies,
 				'header_scripts'  => $header_scripts,
 			]
@@ -289,7 +286,7 @@ class WPSEO_Admin_Asset_Manager {
 		$external_scripts = $this->load_generated_asset_file(
 			[
 				'asset_file'      => __DIR__ . '/../src/generated/assets/externals.php',
-				'ext_length'      => $ext_length,
+				'ext_length'      => 3,
 				'suffix'          => '-package',
 				'base_dir'        => 'externals/',
 				'additional_deps' => $additional_dependencies,
@@ -299,7 +296,7 @@ class WPSEO_Admin_Asset_Manager {
 		$language_scripts = $this->load_generated_asset_file(
 			[
 				'asset_file'      => __DIR__ . '/../src/generated/assets/languages.php',
-				'ext_length'      => $ext_length,
+				'ext_length'      => 3,
 				'suffix'          => '-language',
 				'base_dir'        => 'languages/',
 				'additional_deps' => $additional_dependencies,
@@ -317,6 +314,21 @@ class WPSEO_Admin_Asset_Manager {
 			$renamed_scripts
 		);
 
+		$scripts['installation-success'] = [
+			'name'    => 'installation-success',
+			'src'     => 'installation-success.js',
+			'deps'    => [
+				'wp-a11y',
+				'wp-dom-ready',
+				'wp-components',
+				'wp-element',
+				'wp-i18n',
+				self::PREFIX . 'yoast-components',
+				self::PREFIX . 'externals-components',
+			],
+			'version' => $scripts['installation-success']['version'],
+		];
+
 		$scripts['post-edit-classic'] = [
 			'name'      => 'post-edit-classic',
 			'src'       => $scripts['post-edit']['src'],
@@ -330,12 +342,13 @@ class WPSEO_Admin_Asset_Manager {
 				$scripts['post-edit']['deps']
 			),
 			'in_footer' => ! in_array( 'post-edit-classic', $header_scripts, true ),
+			'version'   => $scripts['post-edit']['version'],
 		];
 
 		$scripts['workouts'] = [
-			'name' => 'workouts',
-			'src'  => 'workouts-' . $flat_version . '.js',
-			'deps' => [
+			'name'    => 'workouts',
+			'src'     => 'workouts.js',
+			'deps'    => [
 				'clipboard',
 				'lodash',
 				'wp-api-fetch',
@@ -353,6 +366,7 @@ class WPSEO_Admin_Asset_Manager {
 				self::PREFIX . 'react-select',
 				self::PREFIX . 'yoast-components',
 			],
+			'version' => $scripts['workouts']['version'],
 		];
 
 		// Add the current language to every script that requires the analysis package.
@@ -418,6 +432,7 @@ class WPSEO_Admin_Asset_Manager {
 				'src'       => $args['base_dir'] . $file,
 				'deps'      => $deps,
 				'in_footer' => ! in_array( $name, $args['header_scripts'], true ),
+				'version'   => $data['version'],
 			];
 		}
 
@@ -646,6 +661,11 @@ class WPSEO_Admin_Asset_Manager {
 			[
 				'name' => 'workouts',
 				'src'  => 'workouts-' . $flat_version,
+				'deps' => [ self::PREFIX . 'monorepo' ],
+			],
+			[
+				'name' => 'installation-success',
+				'src'  => 'installation-success-' . $flat_version,
 				'deps' => [ self::PREFIX . 'monorepo' ],
 			],
 		];