+++ packages/yoastseo/src/languageProcessing/languages/ja/Researcher.js
@@ -9,17 +9,25 @@ import wordsCharacterCount from "./helpers/wordsCharacterCount";
 import customCountLength from "./helpers/countCharacters";
 import matchTransitionWordsHelper from "./helpers/matchTransitionWords";
 import getContentWords from "./helpers/getContentWords";
-import findMultiWordKeyphraseInPageTitle from "./helpers/findExactMatchMultiWordKeyphraseInTitle";
 
 // All config
 import firstWordExceptions from "./config/firstWordExceptions";
 import functionWords from "./config/functionWords";
 import transitionWords from "./config/transitionWords";
 import topicLength from "./config/topicLength";
+import textLength from "./config/textLength";
+import paragraphLength from "./config/paragraphLength";
+import assessmentApplicability from "./config/assessmentApplicabilityCharacterCount";
+import sentenceLength from "./config/sentenceLength";
+import subheadingsTooLong from "./config/subheadingsTooLong";
+import keyphraseLength from "./config/keyphraseLength";
+import metaDescriptionLength from "./config/metaDescriptionLength";
 
 // All custom researches
 import morphology from "./customResearches/getWordForms";
-import getKeywordDensity from "./customResearches/getKeywordDensity";
+import getKeyphraseLength from "./customResearches/getKeyphraseLength";
+import textLengthResearch from "./customResearches/textLength";
+import findKeywordInPageTitle from "./customResearches/findKeywordInPageTitle";
 
 /**
  * The researches contains all the researches
@@ -36,9 +44,7 @@ export default class Researcher extends AbstractResearcher {
 		// Deletes researches that are not available for languages that we haven't supported yet.
 		delete this.defaultResearches.getFleschReadingScore;
 		delete this.defaultResearches.getPassiveVoiceResult;
-
-		// Adds the Japanese custom research to calculate the keyword density.
-		this.addResearch( "getKeywordDensity", getKeywordDensity );
+		delete this.defaultResearches.keywordCountInUrl;
 
 		Object.assign( this.config, {
 			language: "ja",
@@ -46,6 +52,14 @@ export default class Researcher extends AbstractResearcher {
 			functionWords,
 			transitionWords,
 			topicLength,
+			textLength,
+			paragraphLength,
+			assessmentApplicability,
+			sentenceLength,
+			keyphraseLength,
+			subheadingsTooLong,
+			countCharacters: true,
+			metaDescriptionLength,
 		} );
 
 		Object.assign( this.helpers, {
@@ -56,11 +70,13 @@ export default class Researcher extends AbstractResearcher {
 			wordsCharacterCount,
 			customCountLength,
 			matchTransitionWordsHelper,
-			findMultiWordKeyphraseInPageTitle,
 		} );
 
 		Object.assign( this.defaultResearches, {
 			morphology,
+			keyphraseLength: getKeyphraseLength,
+			wordCountInText: textLengthResearch,
+			findKeywordInPageTitle,
 		} );
 	}
 }