+++ packages/yoastseo/src/languageProcessing/researches/getLinkStatistics.js
@@ -1,5 +1,6 @@
 /** @module analyses/getLinkStatistics */
 
+import processExactMatchRequest from "../helpers/match/processExactMatchRequest";
 import filterWordsFromArray from "../helpers/word/filterWordsFromArray";
 import checkNofollow from "../helpers/link/checkNofollow.js";
 import getWords from "../helpers/word/getWords";
@@ -16,8 +17,9 @@ let functionWords = [];
 
 /**
  * Checks whether the link is pointing at itself.
- * @param {string} anchor The link anchor.
- * @param {string} permalink The permalink of the paper.
+ *
+ * @param {string} anchor       The link anchor.
+ * @param {string} permalink    The permalink of the paper.
  *
  * @returns {boolean} Whether the anchor is pointing at itself.
  */
@@ -34,8 +36,9 @@ const linkToSelf = function( anchor, permalink ) {
 
 /**
  * Filters anchors that are not pointing at itself.
- * @param {Array} anchors An array with all anchors from the paper
- * @param {string} permalink The permalink of the paper.
+ *
+ * @param {Array}   anchors     An array with all anchors from the paper.
+ * @param {string}  permalink   The permalink of the paper.
  *
  * @returns {Array} The array of all anchors that are not pointing at the paper itself.
  */
@@ -53,15 +56,18 @@ const filterAnchorsLinkingToSelf = function( anchors, permalink ) {
 
 /**
  * Filters anchors that contain keyphrase or synonyms.
- * @param {Array} anchors An array with all anchors from the paper
- * @param {Object} topicForms The object with topicForms.
- * @param {string} locale The locale of the paper
+ *
+ * @param {Array}   anchors     An array with all anchors from the paper
+ * @param {Object}  topicForms  The object with topicForms.
+ * @param {string}  locale      The locale of the paper.
+ * @param {function}    matchWordCustomHelper The helper function to match word in text.
+ * @param {object}      isExactMatchRequested   An object containing the keyphrase and information whether the exact match has been requested.
  *
  * @returns {Array} The array of all anchors that contain keyphrase or synonyms.
  */
-const filterAnchorsContainingTopic = function( anchors, topicForms, locale ) {
+const filterAnchorsContainingTopic = function( anchors, topicForms, locale, matchWordCustomHelper, isExactMatchRequested ) {
 	const anchorsContainingKeyphraseOrSynonyms = anchors.map( function( anchor ) {
-		return findKeywordInUrl( anchor, topicForms, locale );
+		return findKeywordInUrl( anchor, topicForms, locale, matchWordCustomHelper, isExactMatchRequested );
 	} );
 	anchors = anchors.filter( function( anchor, index ) {
 		return anchorsContainingKeyphraseOrSynonyms[ index ] === true;
@@ -72,13 +78,19 @@ const filterAnchorsContainingTopic = function( anchors, topicForms, locale ) {
 
 /**
  * Filters anchors that are contained within keyphrase or synonyms.
+ *
  * @param {Array}  anchors    An array with all anchors from the paper.
  * @param {Object} topicForms An object containing word forms of words included in the keyphrase or a synonym.
  * @param {string} locale     The locale of the paper.
+ * @param {object} customHelpers            An object containing custom helpers.
+ * @param {object} isExactMatchRequested    An object containing the keyphrase and information whether the exact match has been requested.
  *
  * @returns {Array} The array of all anchors contained in the keyphrase or synonyms.
  */
-const filterAnchorsContainedInTopic = function( anchors, topicForms, locale ) {
+const filterAnchorsContainedInTopic = function( anchors, topicForms, locale, customHelpers, isExactMatchRequested  ) {
+	const matchWordCustomHelper = customHelpers.matchWordCustomHelper;
+	const getWordsCustomHelper = customHelpers.getWordsCustomHelper;
+
 	// Prepare keyphrase and synonym forms for comparison with anchors.
 	const keyphraseAndSynonymsWords = [ flatten( topicForms.keyphraseForms ) ];
 	const synonymsForms = topicForms.synonymsForms;
@@ -90,7 +102,7 @@ const filterAnchorsContainedInTopic = function( anchors, topicForms, locale ) {
 
 	anchors.forEach( function( currentAnchor ) {
 		// Get single words from the anchor.
-		let anchorWords = uniq( getWords( currentAnchor ) );
+		let anchorWords = uniq( getWordsCustomHelper ? getWordsCustomHelper( currentAnchor ) : getWords( currentAnchor ) );
 
 		// Filter function words out of the anchor text.
 		const filteredAnchorWords = filterWordsFromArray( anchorWords, functionWords );
@@ -98,9 +110,16 @@ const filterAnchorsContainedInTopic = function( anchors, topicForms, locale ) {
 			anchorWords = filteredAnchorWords;
 		}
 
-		// Check if anchorWords are contained in the topic phrase words
+		// Check if the exact match is requested and every content words in the anchor text is included in the keyphrase.
+		if ( isExactMatchRequested.exactMatchRequested &&
+			anchorWords.every( anchorWord => isExactMatchRequested.keyphrase.includes( anchorWord ) ) ) {
+			anchorsContainedInTopic.push( true );
+		}
+
+		// Check if anchorWords are contained in the topic phrase words.
 		for ( let i = 0; i < keyphraseAndSynonymsWords.length; i++ ) {
-			if ( anchorWords.every( anchorWord => matchTextWithArray( anchorWord, keyphraseAndSynonymsWords[ i ], locale ).count > 0 ) ) {
+			const topicForm =  keyphraseAndSynonymsWords[ i ];
+			if ( anchorWords.every( anchorWord => matchTextWithArray( anchorWord, topicForm, locale, matchWordCustomHelper ).count > 0 ) ) {
 				anchorsContainedInTopic.push( true );
 				break;
 			}
@@ -114,7 +133,6 @@ const filterAnchorsContainedInTopic = function( anchors, topicForms, locale ) {
 	return anchors;
 };
 
-
 /**
  * Checks whether or not an anchor contains the passed keyword.
  * @param {Paper} paper The paper to research.
@@ -125,6 +143,10 @@ const filterAnchorsContainedInTopic = function( anchors, topicForms, locale ) {
  * @returns {Object} How many anchors contained the keyphrase or synonyms, what are these anchors
  */
 const keywordInAnchor = function( paper, researcher, anchors, permalink ) {
+	const customHelpers = {
+		matchWordCustomHelper: researcher.getHelper( "matchWordCustomHelper" ),
+		getWordsCustomHelper: researcher.getHelper( "getWordsCustomHelper" ),
+	};
 	const result = { totalKeyword: 0, matchedAnchors: [] };
 
 	const keyword = paper.getKeyword();
@@ -143,14 +165,18 @@ const keywordInAnchor = function( paper, researcher, anchors, permalink ) {
 	const locale = paper.getLocale();
 	const topicForms = researcher.getResearch( "morphology" );
 
+	const doubleQuotes = [ "“", "”", "〝", "〞", "〟", "‟", "„", "\"", "\u300c", "\u300d", "\u300e", "\u300f" ];
+
+	const isExactMatchRequested = processExactMatchRequest( keyword, doubleQuotes );
+
 	// Check if any anchors contain keyphrase or synonyms in them.
-	anchors = filterAnchorsContainingTopic( anchors, topicForms, locale );
+	anchors = filterAnchorsContainingTopic( anchors, topicForms, locale, customHelpers.matchWordCustomHelper, isExactMatchRequested );
 	if ( anchors.length === 0 ) {
 		return result;
 	}
 
 	// Check if content words from the anchors are all within the keyphrase or the synonyms.
-	anchors = filterAnchorsContainedInTopic( anchors, topicForms, locale );
+	anchors = filterAnchorsContainedInTopic( anchors, topicForms, locale, customHelpers, isExactMatchRequested );
 	result.totalKeyword = anchors.length;
 	result.matchedAnchors = anchors;